+++ lib/data/bangla_data.dart
@@ -1,0 +1,28 @@
+///[BanglaData] BanglaData for Convert
+class BanglaData {
+  // Bangla months
+  static List<String> banglaMonths = [
+    'বৈশাখ',
+    'জ্যৈষ্ঠ',
+    'আষাঢ়',
+    'শ্রাবণ',
+    'ভাদ্র',
+    'আশ্বিন',
+    'কার্তিক',
+    'অগ্রহায়ণ',
+    'পৌষ',
+    'মাঘ',
+    'ফাল্গুন',
+    'চৈত্র',
+  ];
+  // Bangla weekdays
+  static List<String> banglaWeekdays = [
+    'রবিবার',
+    'সোমবার',
+    'মঙ্গলবার',
+    'বুধবার',
+    'বৃহস্পতিবার',
+    'শুক্রবার',
+    'শনিবার',
+  ];
+}