+++ lib/model/bangla_date_return_model.dart
@@ -1,0 +1,18 @@
+class BanglaDateReturn {
+  final String? date;
+  final String? month;
+  final String? monthDigit;
+  final String? year;
+  final String? yearDescription;
+  final String? weekDay;
+  final String? weekDayDigit;
+  BanglaDateReturn({
+    this.date,
+    this.month,
+    this.monthDigit,
+    this.year,
+    this.yearDescription,
+    this.weekDay,
+    this.weekDayDigit,
+  });
+}